+++ src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/organisms/loginForm";
+import { LoginForm } from "@/features/auth";
 
 export default function LoginPage() {
   return (
